requirements.txt
nltk==3.8.1
torch==2.1.2
openai-whisper==20231117
requests==2.31.0
moviepy==1.0.3
pydub==0.25.1 